Crate tree_sitter_c

This crate provides C language support for the tree-sitter parsing library.

Typically, you will use the language function to add this language to a tree-sitter Parser, and then use the parser to parse some code:

use tree_sitter::Parser;

let code = r#"
int double(int x) {
    return x * 2;
}
"#;
let mut parser = Parser::new();
let language = tree_sitter_c::LANGUAGE;
parser
    .set_language(&language.into())
    .expect("Error loading C parser");
let tree = parser.parse(code, None).unwrap();
assert!(!tree.root_node().has_error());

Constants

HIGHLIGHT_QUERY

The syntax highlighting query for this language.

LANGUAGE

The tree-sitter LanguageFn for this grammar.

NODE_TYPES

The content of the node-types.json file for this grammar.

TAGS_QUERY

The symbol tagging query for this language.